Introduce protocol_redis.h and move the generic definitions in there
